feat: (2단계) 레이아웃 적용하여 회원가입 화면 구현
