+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Signup"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.signup.SignUpActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Signup">
